+++ config/webpack.base.config.js
@@ -2,7 +2,7 @@
  * @Author: junjie.lean
  * @Date: 2019-12-19 13:33:20
  * @Last Modified by: junjie.lean
- * @Last Modified time: 2020-07-29 15:55:02
+ * @Last Modified time: 2020-12-23 17:00:13
  */
 
 /**