+++ src/hooks/useKeyPress.js
@@ -2,22 +2,24 @@
  * @Author: junjie.lean
  * @Date: 2020-01-09 16:09:35
  * @Last Modified by: junjie.lean
- * @Last Modified time: 2020-01-09 17:06:13
+ * @Last Modified time: 2020-12-21 16:17:01
  */
-
+/**
+ * @description 按键监控
+ */
 import React, { useEffect, useState } from "react";
 
 export default function useKeyPress(keyCode) {
   const [isCheckedPass, setCheckedPass] = useState(false);
 
-  const keydownHandle = e => {
+  const keydownHandle = (e) => {
     if (e.keyCode === keyCode) {
       console.log("key down at:", e.keyCode);
       setCheckedPass(true);
     }
   };
 
-  const keyupHandle = e => {
+  const keyupHandle = (e) => {
     if (e.keyCode === keyCode) {
       setCheckedPass(false);
     }