+++ src/index.js
@@ -2,20 +2,25 @@
  * @Author: junjie.lean
  * @Date: 2020-01-09 14:17:37
  * @Last Modified by: junjie.lean
- * @Last Modified time: 2020-07-20 10:11:30
+ * @Last Modified time: 2020-12-23 15:35:52
  */
 
-import "core-js";
+import "core-js/stable";
+import "regenerator-runtime/runtime";
 import React from "react";
-import { render } from "react-dom";
+import ReactDOM from "react-dom";
 import App from "./view/public/public-provider";
 // import svgs from "!!raw-loader!./media/svg/symbol-defs.svg";
-const DOM = document.getElementById("app");
 document.title = "jf-web-app-hook";
 // document.querySelector("#svg").innerHTML += svgs;
 
+function render(props = {}) {
+  const DOM = document.getElementById("app");
+  ReactDOM.render(<App {...props} />, DOM);
+}
+
 if (!window.__POWERED_BY_QIANKUN__) {
-  render(<App />, DOM);
+  render();
 }
 
 export async function bootstrap() {
@@ -26,17 +31,17 @@ export async function bootstrap() {
  * 应用每次进入都会调用 mount 方法，在这里触发应用的渲染方法
  */
 export async function mount(props) {
-  props.onGlobalStateChange((next, prev) => {
-    const nextProps = { ...props, ...next };
-    render(<App {...nextProps} />, DOM);
+  props.onGlobalStateChange((state) => {
+    render(props);
   });
-  render(<App {...props} />, DOM);
+  render(props);
 }
 
 /**
  * 应用每次 切出/卸载 会调用的方法，通常在这里我们会卸载微应用的应用实例
  */
 export async function unmount() {
+  const DOM = document.getElementById("app");
   ReactDOM.unmountComponentAtNode(DOM);
 }